stream_lit_app/requirements.txt
streamlit
pinecone-client==3.1.0
openai==0.27.7
PyMuPDF
tqdm
